ref: added missing fields to entities models feat: implemeneted get_payment method

+++ migrations/2024-06-05-111853_structure/up.sql
@@ -19,7 +19,7 @@ CREATE TABLE payments (
 
 CREATE TABLE charge (
   id INTEGER NOT NULL PRIMARY KEY,
-  amount REAL NOT NULL,
+  amount INTEGER NOT NULL,
   created INTEGER NOT NULL,
   payment_id INTEGER NOT NULL REFERENCES payments
 );
